chore: removed BaseComponentProps and created base in types
